requirements.txt
# Enhanced OSINT Pipeline v2.0 Requirements
# Based on 2024-2025 Industry Best Practices

# Core dependencies
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
Flask>=2.3.0

# Data validation and processing
validators>=0.22.0
dnspython>=2.4.0
python-whois>=0.8.0

# HTTP and API handling
requests>=2.31.0
aiohttp>=3.9.0

# Data analysis and manipulation
pandas>=2.1.0
numpy>=1.24.0

# JSON and data handling
ujson>=5.8.0
orjson>=3.9.0

# Logging and monitoring
colorama>=0.4.6
rich>=13.7.0

# Date and time handling
python-dateutil>=2.8.2
pytz>=2023.3

# Type hints and validation
typing-extensions>=4.8.0
pydantic>=2.5.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Performance and optimization
ujson>=5.8.0
orjson>=3.9.0

# Security and encryption
cryptography>=41.0.0
pyOpenSSL>=23.3.0

# Data export and reporting
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning (for future enhancements)
scikit-learn>=1.3.0
scipy>=1.11.0

# Web scraping (for future enhancements)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Async support
asyncio-mqtt>=0.16.0
aiofiles>=23.2.0

# Configuration management
pyyaml>=6.0.1
toml>=0.10.2

# Progress bars and CLI
tqdm>=4.66.0
click>=8.1.0

# Data serialization
msgpack>=1.0.7

# OSINT and account discovery
holehe>=1.4.0

# Network utilities
netaddr>=0.8.0
ipaddress>=1.0.23

# Text processing
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0

# Database utilities
sqlalchemy>=2.0.0
alembic>=1.12.0

# Caching and performance
redis>=5.0.0
cachetools>=5.3.0

# Monitoring and metrics
prometheus-client>=0.19.0
statsd>=4.0.1

# Error handling and debugging
sentry-sdk>=1.38.0
better-exceptions>=0.3.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
isort>=5.12.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Version control
gitpython>=3.1.0

# Environment management
virtualenv>=20.24.0
pip-tools>=7.3.0
